extract leadership package in its own repository and update Godeps
