+++ Godeps/_workspace/src/github.com/docker/leadership/candidate.go
