+++ Godeps/_workspace/src/github.com/docker/leadership/follower.go
